Struct google_documentai1_beta2::api::GoogleCloudDocumentaiV1beta2Document

pub struct GoogleCloudDocumentaiV1beta2Document {
    pub content: Option<String>,
    pub entities: Option<Vec<GoogleCloudDocumentaiV1beta2DocumentEntity>>,
    pub entity_relations: Option<Vec<GoogleCloudDocumentaiV1beta2DocumentEntityRelation>>,
    pub error: Option<GoogleRpcStatus>,
    pub labels: Option<Vec<GoogleCloudDocumentaiV1beta2DocumentLabel>>,
    pub mime_type: Option<String>,
    pub pages: Option<Vec<GoogleCloudDocumentaiV1beta2DocumentPage>>,
    pub revisions: Option<Vec<GoogleCloudDocumentaiV1beta2DocumentRevision>>,
    pub shard_info: Option<GoogleCloudDocumentaiV1beta2DocumentShardInfo>,
    pub text: Option<String>,
    pub text_changes: Option<Vec<GoogleCloudDocumentaiV1beta2DocumentTextChange>>,
    pub text_styles: Option<Vec<GoogleCloudDocumentaiV1beta2DocumentStyle>>,
    pub uri: Option<String>,
}

Document represents the canonical document resource in Document Understanding AI. It is an interchange format that provides insights into documents and allows for collaboration between users and Document Understanding AI to iterate and optimize for quality.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

documents process projects (response)

locations documents process projects (response)

Fields
content: Option<String>

Optional. Inline document content, represented as a stream of bytes. Note: As with all bytes fields, protobuffers use a pure binary representation, whereas JSON representations use base64.

entities: Option<Vec<GoogleCloudDocumentaiV1beta2DocumentEntity>>

A list of entities detected on Document.text. For document shards, entities in this list may cross shard boundaries.

entity_relations: Option<Vec<GoogleCloudDocumentaiV1beta2DocumentEntityRelation>>

Relationship among Document.entities.

error: Option<GoogleRpcStatus>

Any error that occurred while processing this document.

labels: Option<Vec<GoogleCloudDocumentaiV1beta2DocumentLabel>>

Labels for this document.

mime_type: Option<String>

An IANA published MIME type (also referred to as media type). For more information, see https://www.iana.org/assignments/media-types/media-types.xhtml.

pages: Option<Vec<GoogleCloudDocumentaiV1beta2DocumentPage>>

Visual page layout for the Document.

revisions: Option<Vec<GoogleCloudDocumentaiV1beta2DocumentRevision>>

Revision history of this document.

shard_info: Option<GoogleCloudDocumentaiV1beta2DocumentShardInfo>

Information about the sharding if this document is sharded part of a larger document. If the document is not sharded, this message is not specified.

text: Option<String>

Optional. UTF-8 encoded text in reading order from the document.

text_changes: Option<Vec<GoogleCloudDocumentaiV1beta2DocumentTextChange>>

A list of text corrections made to [Document.text]. This is usually used for annotating corrections to OCR mistakes. Text changes for a given revision may not overlap with each other.

text_styles: Option<Vec<GoogleCloudDocumentaiV1beta2DocumentStyle>>

Styles for the Document.text.

uri: Option<String>

Optional. Currently supports Google Cloud Storage URI of the form gs://bucket_name/object_name. Object versioning is not supported. See Google Cloud Storage Request URIs for more info.